Refactor little

+++ philo_bonus/src/simulation/simulation.c
@@ -6,7 +6,7 @@
 /*   By: susami <susami@student.42tokyo.jp>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/24 15:06:19 by susami            #+#    #+#             */
-/*   Updated: 2022/10/24 23:20:39 by susami           ###   ########.fr       */
+/*   Updated: 2022/10/25 17:31:55 by susami           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -35,10 +35,11 @@ void	start_simulation(t_env *e)
 		if (child_pid < 0)
 			err_exit("fork()");
 		else if (child_pid == 0)
+		{
 			exit(philo_simulation(&e->philosophers[i]));
+		}
 		else
 			e->philosophers[i].pid = child_pid;
-		printf("forked child_pid: %d\n", child_pid);
 		i++;
 	}
 }
@@ -48,6 +49,7 @@ void	wait_simulation_ends(t_env *e)
 	int		status;
 	pid_t	child_pid;
 
+	status = 0;
 	while (true)
 	{
 		child_pid = waitpid(ANY_CHILD_PROCESS, &status, NO_OPTION);
@@ -86,7 +88,10 @@ static void	kill_philosophers_processes(t_env *e)
 	i = 0;
 	while (i < e->args.num_philo)
 	{
-		kill(e->philosophers[i].pid, SIGQUIT);
+		if (e->philosophers[i].pid > 0)
+		{
+			kill(e->philosophers[i].pid, SIGQUIT);
+		}
 		i++;
 	}
 }
